requirements.txt
astroid==1.6.5
atomicwrites==1.3.0
attrs==19.1.0
backports.functools-lru-cache==1.5
bleach==3.1.0
boto3==1.9.120
botocore==1.12.120
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==3.7.3
contextlib2==0.5.5
coverage==4.5.3
DateTime==4.3
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
html5lib==1.0.1
idna==2.8
isort==4.3.15
jellyfish==0.6.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
more-itertools==5.0.0
pathlib2==2.3.3
pbr==5.1.3
pluggy==0.9.0
py==1.8.0
pyaml==18.11.0
pylint==1.9.4
pytest==4.3.1
pytest-cov==2.6.1
python-dateutil==2.8.0
python-lambda==3.2.6
pytz==2018.9
PyYAML==5.1
requests==2.21.0
s3transfer==0.2.0
scandir==1.10.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.3
vcrpy==2.0.1
webencodings==0.5.1
wrapt==1.11.1
zope.interface==4.6.0